Move esbuild into a bundler directory
